==> 1_face_non/R_analysing/data/analysis/6. baseline_diff.Rmd <==
---
title: "baseline_diffs_testing"
author: "Sahar Fatima"
date: "04/06/2020"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Load packages
```{r message=FALSE, warning=FALSE}
library(tidyverse)
library(here)
library(naniar)
library(ggbeeswarm)
library(ggeasy)
library(skimr)

```

# Read in clean data


```{r}

faces <- read_csv(here("R_cleaning", "data", "clean_data", "faces_clean.csv"))

non <- read_csv(here("R_cleaning", "data", "clean_data", "non-faces_clean.csv"))

```

# create participant and condition columns out of filename

Use separate() to split the filename into pp_no and condition and mutate with str_sub to drop the last 6 characters off the end of the values in the condition column to get rid of the .xlsx

```{r}

non_cond <- non %>%
  separate(file_name, into = c("pp_no", "condition"), sep = "_") %>% 
  mutate(condition = str_sub(condition, end = -6)) 
```


# Q1 are there baseline differences between brow and cheek?

1. Filter the data to only include bin 0 (baseline- BL) and plot BL scores for each trial, separately for brow and cheek, coloured by positive/negative. 

```{r}
non_BL <- non_cond %>%
  filter(bin == "bin_0")
```

# choose subset of data from one participant to test code

>Notes from Jenny: the [ ] are one way to select a subset of data, you could also use filter()

> p4810_BL <- non_BL %>%
   filter(pp_no == "pp4810") %>% ....

```{r paged.print=TRUE}
pp4810_BL <- non_BL %>%
  filter(pp_no == "pp4810") %>%
  group_by(muscle, condition) %>%
  mutate(meanRMS = mean(rms), 
         sdRMS = sd(rms))
```


# make new outlier variable 
if sdRMS > 2 = TRUE, if not = FALSE -> outlier2SD
if sdRMS > 1.5 = TRUE, if not = FALSE -> outlier1.5SD


```{r paged.print=TRUE}
pp4810_BL_outlier <- pp4810_BL %>%
  mutate("Doutlier2SD" = if_else(rms > (meanRMS + 2*sdRMS), "TRUE", "FALSE")) %>%
  mutate("outlier1.5SD" = if_else(rms > (meanRMS + 1.5*sdRMS), "TRUE", "FALSE"))
```

It worked! 

# testing it on non-faces data frame 

```{r}
non_BL_stats <- non_BL %>%
  group_by(pp_no, muscle, condition) %>%
  mutate(meanRMS = mean(rms), 
         sdRMS = sd(rms))

non_BL_outlier <- non_BL_stats %>%
  mutate("outlier2SD" = if_else(rms > (meanRMS + 2*sdRMS), "TRUE", "FALSE")) %>%
  mutate("outlier1.5SD" = if_else(rms > (meanRMS + 1.5*sdRMS), "TRUE", "FALSE"))

#check if outlier is logical
glimpse(non_BL_outlier)

#make outlier logical
non_BL_outlier$outlier2SD <- as.logical(non_BL_outlier$outlier2SD)
non_BL_outlier$outlier1.5SD <- as.logical(non_BL_outlier$outlier1.5SD)

```

# testing it on face data frame 

```{r}
face_cond <- faces %>%
  separate(file_name, into = c("pp_no", "condition"), sep = "_") %>% 
  mutate(condition = str_sub(condition, end = -6)) 

face_BL <- face_cond %>%
  filter(bin == "bin_0")

face_BL_stats <- face_BL %>%
  group_by(pp_no, muscle, condition) %>%
  mutate(meanRMS = mean(rms), 
         sdRMS = sd(rms))

face_BL_outlier <- face_BL_stats %>%
  mutate("outlier2SD" = if_else(rms > (meanRMS + 2*sdRMS), "TRUE", "FALSE")) %>%
  mutate("outlier1.5SD" = if_else(rms > (meanRMS + 1.5*sdRMS), "TRUE", "FALSE"))

#check if outlier is logical
glimpse(face_BL_outlier)

#make outlier logical

face_BL_outlier$outlier2SD <- as.logical(face_BL_outlier$outlier2SD)
face_BL_outlier$outlier1.5SD <- as.logical(face_BL_outlier$outlier1.5SD)

```

# write to csv

```{r}
write_csv(non_BL_outlier, here("R_analysing" ,"data", "6_non-face_BL_outlier.csv"))

write_csv(face_BL_outlier, here("R_analysing", "data", "6_face_BL_outlier.csv"))
```

# face -> find problematic trials for each pp and condition 

Plot 429 - face
```{r}
face_BL_outlier %>%
  filter(condition == "429") %>%
  ggplot(aes(x = trial, y = rms, colour = outlier1.5SD)) +
  geom_quasirandom() +
  facet_wrap(~ muscle) +
  labs(title = "face condition 429") + 
  scale_y_continuous(limits = c(0,50))
```

Plot 673 -face
```{r}
face_BL_outlier %>%
  filter(condition == "673") %>%
  ggplot(aes(x = trial, y = rms, colour = outlier1.5SD)) +
  geom_quasirandom() +
  facet_wrap(~ muscle) +
  labs(title = "face condition 673") + 
  scale_y_continuous(limits = c(0,50))
```


Logical variable code as 1 = true 0 = false, so you can sum them. 

```{r}
out_count_face <- face_BL_outlier %>%
  group_by(pp_no, condition, muscle) %>%
  summarise(out_count = sum(outlier1.5SD), trial_count = n())
```


# non-face -> find problematic trials for each pp and condition 

Plot 285 - non-face
```{r}
non_BL_outlier %>%
  filter(condition == "285") %>%
  ggplot(aes(x = trial, y = rms, colour = outlier1.5SD)) +
  geom_quasirandom() +
  facet_wrap(~ muscle) +
  labs(title = "non-face condition 285") + 
  scale_y_continuous(limits = c(0,50))
```

Plot 913 - non-face
```{r}
non_BL_outlier %>%
  filter(condition == "913") %>%
  ggplot(aes(x = trial, y = rms, colour = outlier1.5SD)) +
  geom_quasirandom() +
  facet_wrap(~ muscle) +
  labs(title = "non-face condition 913") + 
  scale_y_continuous(limits = c(0,50))
```

```{r}
out_count_nonface <- non_BL_outlier %>%
  group_by(pp_no, condition, muscle) %>%
  summarise(out_count = sum(outlier1.5SD), trial_count = n())
```


#### plotting face and non-face plots - FALSE points that are high but not outliers

##### face 
condition 429
```{r}
con429_false_highrms <- face_BL_outlier %>%
  filter(condition == "429") %>%
  filter(outlier1.5SD == "FALSE") %>%
  filter(rms >20)

con429_false_highrms %>%
  ggplot(aes(x = trial, y = rms, colour = pp_no)) +
  geom_quasirandom() +
  facet_wrap(~ muscle) +
  labs(title = "face condition 429") + 
  scale_y_continuous(limits = c(0,50))
```

condition 673
> pp547 has v. high sdRMS that is almost = to their meanRMS

```{r}
con673_false_highrms <- face_BL_outlier %>%
  filter(condition == "673") %>%
  filter(outlier1.5SD == "FALSE") %>%
  filter(rms >20)

con673_false_highrms %>%
  ggplot(aes(x = trial, y = rms, colour = pp_no)) +
  geom_quasirandom() +
  facet_wrap(~ muscle) +
  labs(title = "face condition 473") + 
  scale_y_continuous(limits = c(0,50))
```

> pp5417 - seems to have outliers for each trial - brow

##### non-face
work out what is going on with those particular participants 
condition 285
> pp482 and pp544 have pre high sdRMS
> there isn't much of a diff between outlier2SD and outlier 1.5SD

```{r}
con285_false_highrms <- non_BL_outlier %>%
  filter(condition == "285") %>%
  filter(outlier1.5SD == "FALSE") %>%
  filter(rms >20)

con285_false_highrms %>%
  ggplot(aes(x = trial, y = rms, colour = pp_no)) +
  geom_quasirandom() +
  facet_wrap(~ muscle) +
  labs(title = "non-face condition 285") + 
  scale_y_continuous(limits = c(0,50)) 

```


```{r}
con913_false_highrms <- non_BL_outlier %>%
  filter(condition == "913") %>%
  filter(outlier1.5SD == "FALSE") %>%
  filter(rms >20)

con913_false_highrms %>%
  ggplot(aes(x = trial, y = rms, colour = pp_no)) +
  geom_quasirandom() +
  facet_wrap(~ muscle) +
  labs(title = "non-face condition 913") + 
  scale_y_continuous(limits = c(0,50)) 
```
